+++ R/formatdown-deprecated.R
@@ -1,9 +1,0 @@
-#' Deprecated functions in package formatdown.
-#'
-#' The functions listed below are deprecated. Alternative functions with
-#' similar functionality are mentioned. Help pages for deprecated functions
-#' are available at `help("<function>-deprecated")`.
-#'
-#' @name formatdown-deprecated
-#' @keywords internal
-NULL