+++ R/formatdown-package.R
@@ -1,11 +1,31 @@
 #' @keywords internal
 "_PACKAGE"
 
-#' @importFrom checkmate qassert assert_disjunct assert_choice assert_true
-#' @importFrom checkmate test_class check_class assert
-#' @importFrom data.table copy setDT fifelse fcase :=  %like% %between%
-#' @importFrom settings options_manager stop_if_reserved reset
-#' @importFrom units as_units deparse_unit
-#' @importFrom wrapr stop_if_dot_args %?%
+#' @importFrom checkmate assert
+#' @importFrom checkmate assert_choice
+#' @importFrom checkmate assert_disjunct
+#' @importFrom checkmate assert_true
+#' @importFrom checkmate check_class
+#' @importFrom checkmate qassert
+#' @importFrom checkmate test_class
+#'
+#' @importFrom data.table :=
+#' @importFrom data.table %between%
+#' @importFrom data.table %like%
+#' @importFrom data.table copy
+#' @importFrom data.table fcase
+#' @importFrom data.table fifelse
+#' @importFrom data.table setDT
+#' @importFrom data.table tstrsplit
+#'
+#' @importFrom settings options_manager
+#' @importFrom settings reset
+#' @importFrom settings stop_if_reserved
+#'
+#' @importFrom units deparse_unit
+#' @importFrom units as_units
+#'
+#' @importFrom wrapr %?%
+#' @importFrom wrapr stop_if_dot_args
 #'
 NULL