+++ inst/tinytest/test_formatdown_options.R
@@ -12,7 +12,9 @@ test_formatdown_options <- function() {
   # big_interval = 3      any positive integer
   # small_mark = ""       "thin" or "\\\\,"
   # small_interval = 5    any positive integer
-  # whitespace = "\\\\>"  "\\\\:" or "⁠\\\\ ⁠".
+  # whitespace = "\\\\>"  "\\\\:" or "⁠\\\\ ⁠".,
+  # Z = FALSE,
+  # warn = TRUE
 
   # Store existing settings
   prior_settings <- formatdown_options()
@@ -115,6 +117,31 @@ test_formatdown_options <- function() {
   ans919 <- "$6,022 \\cdot 10^{23}$"
   expect_equal(ans919, format_sci(x))
   formatdown_options(reset = TRUE)
+
+  # Z (atomic number)
+  ans920 <- FALSE
+  expect_equal(ans920, formatdown_options("Z"))
+
+  formatdown_options("Z" = TRUE)
+
+  ans921 <- TRUE
+  expect_equal(ans921, formatdown_options("Z"))
+
+  formatdown_options(Z = FALSE)
+
+
+  # isotope warning
+  ans922 <- TRUE
+  expect_equal(ans922, formatdown_options("warn"))
+
+  formatdown_options("warn" = FALSE)
+
+  ans923 <- FALSE
+  expect_equal(ans923, formatdown_options("warn"))
+
+  formatdown_options("warn" = TRUE)
+
+
 
   # ERRORS  ------------------------
   expect_error(formatdown_options("x"))